build: Avoid creating a layer with the OpenEO zip
The OpenEO zip is about 343 MB currently which means that with the current build approach we end up with a container that just has 343 MB of wasted space since we COPY the zip (resulting in 1 layer) and we extract the zip in a subsequent step (resulting in another layer). By doing the unzip and cleanup in a temporary stage we avoid this bloating.